requirements.txt
fastapi
google-adk
google-auth
requests
uvicorn
